requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic[email]==2.6.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.28
alembic==1.13.1

# AWS
boto3==1.34.0
botocore==1.34.0

# Redis
redis==5.0.1

# HTTP client
httpx==0.27.0

# Background tasks
celery==5.3.6
flower==2.0.1

# Observability
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
prometheus-fastapi-instrumentator==6.1.0

# Testing
pytest==8.1.0
pytest-asyncio==0.23.5
httpx==0.27.0
pytest-cov==5.0.0
